// medienverwaltungweb.client/src/app/api-client/model/movieDTO.ts
/**
 * medienVerwaltungWeb.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MovieDTO { 
    id?: number;
    title?: string | null;
    length?: number;
    releaseYear?: number;
    description?: string | null;
    genre?: string | null;
    actorIDs?: Array<number> | null;
    location?: string | null;
    readonly actorCountDisplay?: string | null;
}