// medienverwaltungweb.client/src/app/api-client/model/musicAlbumDTO.ts
/**
 * medienVerwaltungWeb.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MusicAlbumDTO { 
    id?: number;
    title?: string | null;
    songIdList?: Array<number> | null;
    interpretFullName?: string | null;
    location?: string | null;
}